[FEAT]: UILabel, UIStackView 재사용코드 만들고 ProfileLabelView 적용
