+++ InstaClone_CodeBase/InstaClone_CodeBase/View/Componants/UILabel +.swift
@@ -1,0 +1,12 @@
+import UIKit
+
+extension UILabel {
+    
+    func profileLabel(text: String, textColor: UIColor, fontSize: CGFloat){
+        self.text = text
+        self.translatesAutoresizingMaskIntoConstraints = false
+        self.textColor = textColor
+        self.font = UIFont.systemFont(ofSize: fontSize)
+    }
+    
+}
