+++ InstaClone_CodeBase/InstaClone_CodeBase/View/Componants/UIStackView +.swift
@@ -1,0 +1,18 @@
+import UIKit
+
+extension UIStackView {
+    
+    func verticalStackView(spacing: CGFloat){
+        self.translatesAutoresizingMaskIntoConstraints = false
+        self.axis = .vertical
+        self.alignment = .center
+        self.spacing = spacing
+    }
+    
+    func horizontalStackView(spacing: CGFloat){
+        self.translatesAutoresizingMaskIntoConstraints = false
+        self.axis = .horizontal
+        self.alignment = .center
+        self.spacing = spacing
+    }
+}
